use double instead of float

+++ Assets/Scripts/CombatStats.cs
@@ -7,18 +7,18 @@ public class CombatStats
 {
 
     /// Increases blunt damage
-    public float strength = 1.0;
+    public double strength = 1.0;
     /// Increases piercing damage
-    public float dexterity = 1.0;
+    public double dexterity = 1.0;
     /// Increases intelligence damage
-    public float intelligence = 1.0;
+    public double intelligence = 1.0;
 
     // Decreases blunt damage
-    public float constitution = 1.0;
+    public double constitution = 1.0;
     /// Increases piercing damage
-    public float agility = 1.0;
+    public double agility = 1.0;
     /// Increases intelligence damage
-    public float willpower = 1.0;
+    public double willpower = 1.0;
 
 
     /// Is always created with default values
